refactor: rename promo credit banner files
