+++ src/util/promoCredit.js
